Somachar/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.3
bootstrap4==0.1.0
bs4==0.0.1
certifi==2019.3.9
cffi==1.14.4
chardet==3.0.4
click==7.1.2
comtypes==1.1.7
cryptography==3.3.1
DateTime==4.3
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.44.0
django-bootstrap4==2.3.1
django-compat==1.0.15
django-crispy-forms==1.10.0
django-extensions==3.1.0
django-heroku==0.3.1
django-sslserver==0.22
django-utils-six==2.0
djangorestframework==3.12.2
docopt==0.6.2
filelock==3.0.12
form==0.0.1
gunicorn==20.0.4
html5lib==1.1
idna==2.8
itsdangerous==1.1.0
joblib==0.17.0
Js2Py==0.70
lxml==4.6.1
MarkupSafe==1.1.1
oauthlib==3.0.1
packaging==20.4
pipwin==0.5.0
psycopg2==2.8.6
pycparser==2.20
pyjsparser==2.7.1
PyJWT==2.0.0
pyparsing==2.4.7
PyPrind==2.11.2
pySmartDL==1.3.4
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.3.1
regex==2020.10.28
requests==2.21.0
requests-file==1.5.1
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
sqlparse==0.3.0
tqdm==4.51.0
tzlocal==2.1
urllib3==1.24.3
virtualenv==20.1.0
webencodings==0.5.1
whitenoise==5.2.0
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
zope.interface==5.2.0
